% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fwc.R
\name{read_fwc}
\alias{read_fwc}
\title{Download and import enterprise agreements data from the Fair Work Commission}
\usage{
read_fwc(path = tempdir())
}
\arguments{
\item{path}{Path to directory in which downloaded Excel file should be stored}
}
\description{
Download and import enterprise agreements data from the Fair Work Commission
}
\examples{
read_fwc()

}
